Split SerializedDataHandlingTest and put it in correct directories
